requirements.txt
docopt==0.6.2
python-digitalocean==1.15.0
yaspin==0.16.0
